Create driver_control_code
This is the main file where we'll have all the code we need to have a driver control the robot. Adrian, please upload the code that Patrick made for last year that is stored on the app so I can trim it down to what we need and start coding everything apart from the wheel motors.